requirements.txt
runpod
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
pillow==10.1.0
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git
